Bug修复

+++ Editor/03.处理类型/NetworkDelegateProcess.cs
@@ -84,7 +84,7 @@ namespace JFramework.Editor
             worker.Emit(OpCodes.Ldstr, md.FullName);
             worker.Emit(OpCodes.Ldc_I4, (int)NetworkUtility.GetStableId(md.FullName));
             worker.Emit(OpCodes.Ldloc_0);
-            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<byte>());
+            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<int>());
             worker.Emit(ca.GetFieldType<bool>() ? OpCodes.Ldc_I4_1 : OpCodes.Ldc_I4_0);
             worker.Emit(OpCodes.Callvirt, models.sendClientRpcInternal);
             NetworkBehaviourProcess.WritePushWriter(worker, models);
@@ -160,7 +160,7 @@ namespace JFramework.Editor
             worker.Emit(OpCodes.Ldstr, md.FullName);
             worker.Emit(OpCodes.Ldc_I4, (int)NetworkUtility.GetStableId(md.FullName));
             worker.Emit(OpCodes.Ldloc_0);
-            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<byte>());
+            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<int>());
             worker.Emit(OpCodes.Call, models.sendServerRpcInternal);
             NetworkBehaviourProcess.WritePushWriter(worker, models);
             worker.Emit(OpCodes.Ret);
@@ -235,7 +235,7 @@ namespace JFramework.Editor
             worker.Emit(OpCodes.Ldstr, md.FullName);
             worker.Emit(OpCodes.Ldc_I4, (int)NetworkUtility.GetStableId(md.FullName));
             worker.Emit(OpCodes.Ldloc_0);
-            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<byte>());
+            worker.Emit(OpCodes.Ldc_I4, ca.GetFieldType<int>());
             worker.Emit(OpCodes.Callvirt, models.sendTargetRpcInternal);
             NetworkBehaviourProcess.WritePushWriter(worker, models);
             worker.Emit(OpCodes.Ret);
